// CdT.ClientPortal.Front/angularApp/app/model/breeze/department.ts
import { EntityBase } from './entity-base';
import { Client } from './client';
import { Section } from './section';

/// <code-import> Place custom imports between <code-import> tags

/// </code-import>

export class Department extends EntityBase {

    /// <code> Place custom code between <code> tags
    
    /// </code>

    // Generated code. Do not place code below this line.
    id: string;
    version: number;
    createdBy: string;
    updatedBy: string;
    creationDate: Date;
    updateDate: Date;
    isDeleted: boolean;
    code: string;
    defaultLabel: string;
    validFrom: Date;
    validTo: Date;
    description: string;
    displayOrder: number;
    isUsedInEcdt: boolean;
    abbreviation: string;
    clientId: string;
    clientPortalId: number;
    client: Client;
    sections: Section[];
}